Доработка по ТЗ
После перечитывания ТЗ оказалось, что надо вводить после заданных b своё b тоже.
Был исправлен баг с неточностями тем, что в интерфейсах предлагается float и unsigned, а внутренние вычисления происходят по unsigned long и double.

+++ labs/z7.16_methodsOfApproximateIntegralCalculus/z7.16 test integral Right Rectangle Method.h
@@ -45,7 +45,7 @@ unsigned short z7_16_tiRRM_p_Test_p_getH()
 	else
 	{
 		out++;
-		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_getH: 0, 0: failed. Result: ", z7_16_iRRM_p_getH(0.0f, 0.0f, 0u));
+		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_getH: 0, 0: failed. Result: ", (float)z7_16_iRRM_p_getH(0.0f, 0.0f, 0u));
 	}
 
 	if (z7_16_iRRM_p_getH(-4.0f, 7.2f, 4u) == 2.8f)
@@ -53,14 +53,14 @@ unsigned short z7_16_tiRRM_p_Test_p_getH()
 	else
 	{
 		out++;
-		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_getH: -3, 0.2: failed. Result: ", z7_16_iRRM_p_getH(-4.0f, 7.2f, 4));
+		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_getH: -3, 0.2: failed. Result: ", (float)z7_16_iRRM_p_getH(-4.0f, 7.2f, 4));
 	}
 
 	return out;
 }
 
-float z7_16_tiRRM_p_Test_p_sum_functionArgument1(float x) { return 1.0f; }
-float z7_16_tiRRM_p_Test_p_sum_functionArgument2(float x) { return x; }
+double z7_16_tiRRM_p_Test_p_sum_functionArgument1(double x) { return 1.0f; }
+double z7_16_tiRRM_p_Test_p_sum_functionArgument2(double x) { return x; }
 
 // Тестирует z7_16_iRRM_p_sum. Возвращает количество ошибок.
 unsigned short z7_16_tiRRM_p_Test_p_sum()
@@ -71,7 +71,7 @@ unsigned short z7_16_tiRRM_p_Test_p_sum()
 	else
 	{
 		out++;
-		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_sum: 0, f1, 1.0: failed. Result: ", z7_16_iRRM_p_sum(0, z7_16_tiRRM_p_Test_p_sum_functionArgument1, 1.0f, 0.0f));
+		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_sum: 0, f1, 1.0: failed. Result: ", (float)z7_16_iRRM_p_sum(0, z7_16_tiRRM_p_Test_p_sum_functionArgument1, 1.0f, 0.0f));
 	}
 
 	if (z7_16_iRRM_p_sum(3, z7_16_tiRRM_p_Test_p_sum_functionArgument2, 1.0f, 0.0f) == 3.0f) // Смотреть примечание 1 в коде
@@ -79,7 +79,7 @@ unsigned short z7_16_tiRRM_p_Test_p_sum()
 	else
 	{
 		out++;
-		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_sum:  3, f2, 1.0: failed. Result: ", z7_16_iRRM_p_sum(3, z7_16_tiRRM_p_Test_p_sum_functionArgument2, 1.0f, 0.0f));
+		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_p_sum:  3, f2, 1.0: failed. Result: ", (float)z7_16_iRRM_p_sum(3, z7_16_tiRRM_p_Test_p_sum_functionArgument2, 1.0f, 0.0f));
 	}
 	return out;
 }
@@ -89,7 +89,7 @@ unsigned short z7_16_tiRRM_p_Test_CalculateIntegralByN()
 	unsigned short out = 0;
 	float buffer;
 
-	buffer = z7_16_iRRM_CalculateIntegralByN(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 16.0f, 10000000);
+	buffer = (float)z7_16_iRRM_CalculateIntegralByN(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 16.0f, 10000000);
 	if (buffer - 128.0f < 0.1f && buffer - 128.0f > -0.1f)
 		z7_16_test_UserInterface_p_logStringFloat(TRUE, "z7_16_iRRM_CalculateIntegralByN: f2, 0, 16, 10000000: ok. Observational error: ", buffer - 128.0f);
 	else
@@ -98,7 +98,7 @@ unsigned short z7_16_tiRRM_p_Test_CalculateIntegralByN()
 		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_CalculateIntegralByN: f2, 0, 16, 10000000: failed. Observational error: ", buffer - 128.0f);
 	}
 
-	buffer = z7_16_iRRM_CalculateIntegralByN(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 0.0f, 0);
+	buffer = (float)z7_16_iRRM_CalculateIntegralByN(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 0.0f, 0);
 	if (ISNAN(buffer))
 		z7_16_test_UserInterface_p_log(TRUE, "z7_16_iRRM_CalculateIntegralByN: f2, 0, 0, 0: isnan. ok.");
 	else
@@ -115,7 +115,7 @@ unsigned short z7_16_tiRRM_p_Test_CalculateIntegralByH()
 	unsigned short out = 0;
 	float buffer;
 
-	buffer = z7_16_iRRM_CalculateIntegralByH(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 16.0f, 0.01f);
+	buffer = (float)z7_16_iRRM_CalculateIntegralByH(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 16.0f, 0.01f);
 	if (buffer - 128.0f < 0.01f && buffer - 128.0f > -0.01f)
 		z7_16_test_UserInterface_p_logStringFloat(TRUE, "z7_16_iRRM_CalculateIntegralByH: f2, 0, 16, 0.01: ok. Observational error: ", buffer - 128.0f);
 	else
@@ -124,7 +124,7 @@ unsigned short z7_16_tiRRM_p_Test_CalculateIntegralByH()
 		z7_16_test_UserInterface_p_logStringFloat(FALSE, "z7_16_iRRM_CalculateIntegralByH: f2, 0, 16, 10000000: failed. Observational error: ", buffer - 128.0f);
 	}
 
-	buffer = z7_16_iRRM_CalculateIntegralByH(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 0.0f, 0.0f);
+	buffer = (float)z7_16_iRRM_CalculateIntegralByH(z7_16_tiRRM_p_Test_p_sum_functionArgument2, 0.0f, 0.0f, 0.0f);
 	if (ISNAN(buffer))
 		z7_16_test_UserInterface_p_log(TRUE, "z7_16_iRRM_CalculateIntegralByH: f2, 0, 0, 0: isnan. ok.");
 	else
